Add support for new instance type needed for cloud.
Summary:
Since we want to use 0.3 cores for our cloud deployments, we need to support float values
for number of cores in our instance types. We also need to add the required kind of instance to our
DB.

Test Plan:
Created a k8s provider and verified that the instance type got populated with a floating
point value. Also tested AWS provider create and verified that nothing changed in the instance
creations for that.

Reviewers: ram

Reviewed By: ram

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D7907

+++ managed/src/main/java/com/yugabyte/yw/cloud/UniverseResourceDetails.java
@@ -36,7 +36,7 @@ public class UniverseResourceDetails {
 
   public double pricePerHour = 0;
   public double ebsPricePerHour = 0;
-  public int numCores = 0;
+  public double numCores = 0;
   public double memSizeGB = 0;
   public int volumeCount = 0;
   public int volumeSizeGB = 0;
@@ -67,7 +67,7 @@ public class UniverseResourceDetails {
     azList.add(azName);
   }
 
-  public void addNumCores(int cores) {
+  public void addNumCores(double cores) {
     numCores += cores;
   }
